Moves local resource validation code into a separate file and export
This was part of an attempt to validate local resources before deploy
for the controller side deploy (to be able to call these from the
application/deployer). The validation part didn't pan out, but this
might be a nice tidy up for that validation code.
